GEOMESA-692 Remove Duplicated Interface
* AccumuloBackedRasterOperations has now been replaced by the RasterStore
